restructure modules into directories
