+++ src/dispatcher/mod.rs
@@ -1,0 +1,3 @@
+mod core;
+
+pub use core::{Dispatcher, HandlerRequest, HandlerResponse, HandlerSender};
