+++ src/generator/core.rs
@@ -435,7 +435,7 @@ pub fn is_named_type(ty: &str) -> bool {
     !primitives.contains(&ty) && matches!(ty.chars().next(), Some('A'..='Z'))
 }
 
-fn unique_handler_name(seen: &mut HashSet<String>, name: &str) -> String {
+pub(crate) fn unique_handler_name(seen: &mut HashSet<String>, name: &str) -> String {
     if !seen.contains(name) {
         seen.insert(name.to_string());
         return name.to_string();
@@ -456,21 +456,6 @@ fn unique_handler_name(seen: &mut HashSet<String>, name: &str) -> String {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_unique_handler_name() {
-        let mut seen = HashSet::new();
-        let a = unique_handler_name(&mut seen, "foo");
-        assert_eq!(a, "foo");
-        let b = unique_handler_name(&mut seen, "foo");
-        assert_eq!(b, "foo_1");
-        let c = unique_handler_name(&mut seen, "foo");
-        assert_eq!(c, "foo_2");
-    }
-}
 
 pub fn rust_literal_for_example(field: &FieldDef, example: &Value) -> String {
     let literal = match example {
