+++ src/router/mod.rs
@@ -1,0 +1,5 @@
+mod core;
+#[cfg(test)]
+mod tests;
+
+pub use core::{RouteMatch, Router};
