+++ src/server/mod.rs
@@ -1,0 +1,3 @@
+mod core;
+
+pub use core::AppService;
