+++ src/spec/core.rs
